fix(lemma-finance): Cleanup
