+++ src/apps/lemma-finance/optimism/lemma-finance.x-synth.token-fetcher.ts
@@ -59,8 +59,8 @@ export class OptimismLemmaFinanceXSynthTokenFetcher implements PositionFetcher<A
 
   async getPositions() {
     const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
-
     const multicall = this.appToolkit.getMulticall(network);
+
     const tokens = await Promise.all(
       xSynthAddresses.map(async position => {
         const contract = this.contractFactory.xSynth({
