+++ src/apps/lemma-finance/optimism/lemma-finance.x-usdl.token-fetcher.ts
@@ -11,14 +11,14 @@ import { AppTokenPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
 import { LemmaFinanceContractFactory } from '../contracts';
-import LEMMA_FINANCE_DEFINITION from '../lemma-finance.definition';
+import { LEMMA_FINANCE_DEFINITION } from '../lemma-finance.definition';
 
 const appId = LEMMA_FINANCE_DEFINITION.id;
-const groupId = LEMMA_FINANCE_DEFINITION.groups.usdl.id;
+const groupId = LEMMA_FINANCE_DEFINITION.groups.xUsdl.id;
 const network = Network.OPTIMISM_MAINNET;
 
 @Register.TokenPositionFetcher({ appId, groupId, network })
-export class OptimismLemmaFinanceUsdlTokenFetcher implements PositionFetcher<AppTokenPosition> {
+export class OptimismLemmaFinanceXUsdlTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(
     @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
     @Inject(LemmaFinanceContractFactory) private readonly contractFactory: LemmaFinanceContractFactory,
@@ -34,19 +34,19 @@ export class OptimismLemmaFinanceUsdlTokenFetcher implements PositionFetcher<App
       '0x76fb31fb4af56892a25e32cfc43de717950c9278', // AAVE
     ];
 
-    const usdlAddress = '0x96f2539d3684dbde8b3242a51a73b66360a5b541';
-
+    const xusdlAddress = '0x252ea7e68a27390ce0d53851192839a39ab8b38c';
     const multicall = this.appToolkit.getMulticall(network);
-    const contract = this.contractFactory.usdl({
-      address: usdlAddress,
+    const contract = this.contractFactory.xUsdl({
+      address: xusdlAddress,
       network,
     });
 
-    const [name, symbol, decimals, supplyRaw] = await Promise.all([
+    const [name, symbol, decimals, supplyRaw, assetsPerShare] = await Promise.all([
       multicall.wrap(contract).name(),
       multicall.wrap(contract).symbol(),
       multicall.wrap(contract).decimals(),
       multicall.wrap(contract).totalSupply(),
+      multicall.wrap(contract).assetsPerShare(),
     ]);
 
     const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
@@ -61,18 +61,18 @@ export class OptimismLemmaFinanceUsdlTokenFetcher implements PositionFetcher<App
     const tokens = _.compact(tokensRaw);
 
     const supply = Number(supplyRaw) / 10 ** decimals;
-    const price = 1;
+    const price = Number(assetsPerShare) / 10 ** decimals;
     const pricePerShare = 1;
 
     const label = `${name} (${symbol})`;
-    const images = [getAppAssetImage(appId, 'USDL')];
+    const images = [getAppAssetImage(appId, 'xUSDL')];
     const secondaryLabel = buildDollarDisplayItem(price);
 
     const token: AppTokenPosition = {
       type: ContractType.APP_TOKEN,
       appId,
       groupId,
-      address: usdlAddress,
+      address: xusdlAddress,
       network,
       symbol,
       decimals,
